screen apply macid
