+++ src/Components/Screen/Screens.jsx
@@ -148,10 +148,12 @@ const Screens = ({ sidebarOpen, setSidebarOpen }) => {
   const [tagUpdateScreeen, setTagUpdateScreeen] = useState(null);
   const [selectedYoutube, setSelectedYoutube] = useState();
   const [selectedTextScroll, setSelectedTextScroll] = useState();
-
+  const [setscreenMacID, setSetscreenMacID] = useState("");
   const { loading, screens, deleteLoading } = useSelector((s) => s.root.screen);
   const { schedules } = useSelector((s) => s.root.schedule);
   const { compositions } = useSelector((s) => s.root.composition);
+
+  console.log("screens", screens);
 
   const dispatch = useDispatch();
 
@@ -575,67 +577,67 @@ const Screens = ({ sidebarOpen, setSidebarOpen }) => {
     }
   };
 
-  // for call signal R
-  useEffect(() => {
-    const connectSignalR = async () => {
-      console.log("run signal r");
-      const newConnection = new HubConnectionBuilder()
-        .withUrl(SIGNAL_R)
-        .configureLogging(LogLevel.Information)
-        .withAutomaticReconnect()
-        .build();
+  // // for call signal R
+  // useEffect(() => {
+  //   const connectSignalR = async () => {
+  //     console.log("run signal r");
+  //     const newConnection = new HubConnectionBuilder()
+  //       .withUrl(SIGNAL_R)
+  //       .configureLogging(LogLevel.Information)
+  //       .withAutomaticReconnect()
+  //       .build();
 
-      newConnection.on("ScreenConnected", (screenConnected) => {
-        console.log("ScreenConnected", screenConnected);
-        setScreenConnected(screenConnected);
-      });
+  //     newConnection.on("ScreenConnected", (screenConnected) => {
+  //       console.log("ScreenConnected", screenConnected);
+  //       setScreenConnected(screenConnected);
+  //     });
 
-      newConnection.on("TvStatus", (UserID, ScreenID, status) => {
-        console.log("SendTvStatus", UserID, ScreenID, status);
-        setSendTvStatus(status);
-        setSendTvStatusScreenID(ScreenID);
-        // debugger;
-        var b = document.getElementById("changetvstatus" + ScreenID);
-        b.setAttribute(
-          "class",
-          "rounded-full px-6 py-2 text-white text-center " +
-            (status == true ? "bg-[#3AB700]" : "bg-[#FF0000]")
-        );
-        b.textContent = status == true ? "Live" : "offline";
-      });
+  //     newConnection.on("TvStatus", (UserID, ScreenID, status) => {
+  //       console.log("SendTvStatus", UserID, ScreenID, status);
+  //       setSendTvStatus(status);
+  //       setSendTvStatusScreenID(ScreenID);
+  //       // debugger;
+  //       var b = document.getElementById("changetvstatus" + ScreenID);
+  //       b.setAttribute(
+  //         "class",
+  //         "rounded-full px-6 py-2 text-white text-center " +
+  //           (status == true ? "bg-[#3AB700]" : "bg-[#FF0000]")
+  //       );
+  //       b.textContent = status == true ? "Live" : "offline";
+  //     });
 
-      try {
-        await newConnection.start();
-        console.log("Connection established");
-        setConnection(newConnection);
+  //     try {
+  //       await newConnection.start();
+  //       console.log("Connection established");
+  //       setConnection(newConnection);
 
-        // Invoke ScreenConnected method
-        await newConnection.invoke("ScreenConnected");
-        console.log("Message sent:", screenConnected);
+  //       // Invoke ScreenConnected method
+  //       await newConnection.invoke("ScreenConnected", "MacID");
+  //       console.log("Message sent:", screenConnected);
 
-        // Invoke SendTvStatus method
-        //await newConnection.invoke("SendTvStatus", true, "E0:76:D0:32:54:00");
-        //console.log("SendTvStatus", 1, "E0:76:D0:32:54:00");
-      } catch (error) {
-        console.error("Error during connection:", error);
-      }
-    };
+  //       // Invoke SendTvStatus method
+  //       //await newConnection.invoke("SendTvStatus", true, "E0:76:D0:32:54:00");
+  //       //console.log("SendTvStatus", 1, "E0:76:D0:32:54:00");
+  //     } catch (error) {
+  //       console.error("Error during connection:", error);
+  //     }
+  //   };
 
-    connectSignalR(); // Call the combined function
+  //   connectSignalR(); // Call the combined function
 
-    return () => {
-      if (connection) {
-        connection
-          .stop()
-          .then(() => {
-            console.log("Connection stopped");
-          })
-          .catch((error) => {
-            console.error("Error stopping connection:", error);
-          });
-      }
-    };
-  }, []);
+  //   return () => {
+  //     if (connection) {
+  //       connection
+  //         .stop()
+  //         .then(() => {
+  //           console.log("Connection stopped");
+  //         })
+  //         .catch((error) => {
+  //           console.error("Error stopping connection:", error);
+  //         });
+  //     }
+  //   };
+  // }, []);
 
   // useEffect(() => {
   //   if (connection) {
@@ -1285,7 +1287,9 @@ const Screens = ({ sidebarOpen, setSidebarOpen }) => {
                               <Link
                                 to={`/screensplayer?screenID=${screen.screenID}`}
                               >
-                                {screen?.screenName?.length > 10 ? screen?.screenName.slice(0, 10) + "..." : screen.screenName}
+                                {screen?.screenName?.length > 10
+                                  ? screen?.screenName.slice(0, 10) + "..."
+                                  : screen.screenName}
                               </Link>
                               <button
                                 onClick={() => {
@@ -1305,7 +1309,6 @@ const Screens = ({ sidebarOpen, setSidebarOpen }) => {
                           {screen.googleLocation}
                         </td>
                       )}
-
 
                       {statusContentVisible && (
                         <td className="p-2 text-center">
@@ -1334,6 +1337,7 @@ const Screens = ({ sidebarOpen, setSidebarOpen }) => {
                           <div
                             onClick={(e) => {
                               setAssetScreenID(screen.screenID);
+                              setSetscreenMacID(screen.macid);
                               setShowAssetModal(true);
                               setSelectedAsset({
                                 ...selectedAsset,
@@ -1363,6 +1367,7 @@ const Screens = ({ sidebarOpen, setSidebarOpen }) => {
                               selectedTextScroll={selectedTextScroll}
                               selectedYoutube={selectedYoutube}
                               selectedAsset={selectedAsset}
+                              setscreenMacID={setscreenMacID}
                             />
                           )}
                         </td>
@@ -1531,9 +1536,10 @@ const Screens = ({ sidebarOpen, setSidebarOpen }) => {
                       )}
 
                       {tagsContentVisible && (
-                        <td 
-                        title={screen?.tags && screen?.tags}
-                        className="p-2 text-center flex flex-wrap items-center justify-center gap-2 mt-6 break-all">
+                        <td
+                          title={screen?.tags && screen?.tags}
+                          className="p-2 text-center flex flex-wrap items-center justify-center gap-2 mt-6 break-all"
+                        >
                           {(screen?.tags === "" || screen?.tags === null) && (
                             <span>
                               <AiOutlinePlusCircle
@@ -1558,7 +1564,8 @@ const Screens = ({ sidebarOpen, setSidebarOpen }) => {
                                   screen.tags.split(",").length > 2
                                     ? 3
                                     : screen.tags.split(",").length
-                                ).map((text) => {
+                                )
+                                .map((text) => {
                                   if (text.toString().length > 10) {
                                     return text
                                       .split("")
@@ -1642,304 +1649,340 @@ const Screens = ({ sidebarOpen, setSidebarOpen }) => {
               ) : (
                 filteredScreenData.map((screen) => (
                   <tbody key={screen.screenID}>
-                  <tr className="border-b border-b-[#E4E6FF]">
-                    {screenContentVisible && (
-                      <td className="flex items-center ">
-                        <input
-                          type="checkbox"
-                          className="mr-3"
-                          style={{
-                            display: selectAllChecked ? "block" : "none",
-                          }}
-                          onChange={() =>
-                            handleScreenCheckboxChange(screen.screenID)
-                          }
-                          checked={screenCheckboxes[screen.screenID]}
-                        />
+                    <tr className="border-b border-b-[#E4E6FF]">
+                      {screenContentVisible && (
+                        <td className="flex items-center ">
+                          <input
+                            type="checkbox"
+                            className="mr-3"
+                            style={{
+                              display: selectAllChecked ? "block" : "none",
+                            }}
+                            onChange={() =>
+                              handleScreenCheckboxChange(screen.screenID)
+                            }
+                            checked={screenCheckboxes[screen.screenID]}
+                          />
 
-                        {isEditingScreen &&
-                        editingScreenID === screen.screenID ? (
-                          <div className="flex items-center gap-2">
-                            <input
-                              type="text"
-                              className="border border-primary rounded-md w-full"
-                              value={editedScreenName}
-                              onChange={(e) => {
-                                setEditedScreenName(e.target.value);
-                              }}
-                            />
-                            <button
-                              onClick={() => {
-                                handleScreenNameUpdate(screen.screenID);
-                              }}
-                            >
-                              <AiOutlineSave className="text-2xl ml-1 hover:text-primary" />
-                            </button>
-                            {/* {screenNameError && (
+                          {isEditingScreen &&
+                          editingScreenID === screen.screenID ? (
+                            <div className="flex items-center gap-2">
+                              <input
+                                type="text"
+                                className="border border-primary rounded-md w-full"
+                                value={editedScreenName}
+                                onChange={(e) => {
+                                  setEditedScreenName(e.target.value);
+                                }}
+                              />
+                              <button
+                                onClick={() => {
+                                  handleScreenNameUpdate(screen.screenID);
+                                }}
+                              >
+                                <AiOutlineSave className="text-2xl ml-1 hover:text-primary" />
+                              </button>
+                              {/* {screenNameError && (
                               <div className="text-red">
                                 {screenNameError}
                               </div>
                             )} */}
-                          </div>
-                        ) : (
-                          <div className="flex items-center  gap-2">
-                            <Link
-                              to={`/screensplayer?screenID=${screen.screenID}`}
-                            >
-                              {screen?.screenName?.length > 10 ? screen?.screenName.slice(0, 10) + "..." : screen.screenName}
-                            </Link>
-                            <button
-                              onClick={() => {
-                                setIsEditingScreen(true);
-                                setEditingScreenID(screen.screenID);
-                                setEditedScreenName(screen?.screenName);
-                              }}
-                            >
-                              <MdOutlineModeEdit className="w-6 h-6 hover:text-primary" />
-                            </button>
-                          </div>
-                        )}
-                      </td>
-                    )}
-                    {locContentVisible && (
-                      <td className="p-2 break-words text-center">
-                        {screen.googleLocation}
-                      </td>
-                    )}
+                            </div>
+                          ) : (
+                            <div className="flex items-center  gap-2">
+                              <Link
+                                to={`/screensplayer?screenID=${screen.screenID}`}
+                              >
+                                {screen?.screenName?.length > 10
+                                  ? screen?.screenName.slice(0, 10) + "..."
+                                  : screen.screenName}
+                              </Link>
+                              <button
+                                onClick={() => {
+                                  setIsEditingScreen(true);
+                                  setEditingScreenID(screen.screenID);
+                                  setEditedScreenName(screen?.screenName);
+                                }}
+                              >
+                                <MdOutlineModeEdit className="w-6 h-6 hover:text-primary" />
+                              </button>
+                            </div>
+                          )}
+                        </td>
+                      )}
+                      {locContentVisible && (
+                        <td className="p-2 break-words text-center">
+                          {screen.googleLocation}
+                        </td>
+                      )}
 
-
-                    {statusContentVisible && (
-                      <td className="p-2 text-center">
-                        <button
-                          id={`changetvstatus${screen.screenID}`}
-                          className={`rounded-full px-6 py-2 text-white text-center ${
-                            screen.screenStatus == 1
-                              ? "bg-[#3AB700]"
-                              : "bg-[#FF0000]"
-                          }`}
-                        >
-                          {screen.screenStatus == 1 ? "Live" : "offline"}
-                        </button>
-                      </td>
-                    )}
-                    {/* {lastSeenContentVisible && (
+                      {statusContentVisible && (
+                        <td className="p-2 text-center">
+                          <button
+                            id={`changetvstatus${screen.screenID}`}
+                            className={`rounded-full px-6 py-2 text-white text-center ${
+                              screen.screenStatus == 1
+                                ? "bg-[#3AB700]"
+                                : "bg-[#FF0000]"
+                            }`}
+                          >
+                            {screen.screenStatus == 1 ? "Live" : "offline"}
+                          </button>
+                        </td>
+                      )}
+                      {/* {lastSeenContentVisible && (
                       <td className="p-2 text-center break-words">
                         25 May 2023
                       </td>
                     )} */}
-                    {nowPlayingContentVisible && (
-                      <td
-                        className="p-2 text-center "
-                        style={{ wordBreak: "break-all" }}
-                      >
-                        <div
-                          onClick={(e) => {
-                            setAssetScreenID(screen.screenID);
-                            setShowAssetModal(true);
-                            setSelectedAsset({
-                              ...selectedAsset,
-                              assetName: e.target.value,
-                            });
-                            setSelectedAsset(screen?.assetName);
-                          }}
-                          title={screen?.assetName}
-                          className="flex items-center justify-between gap-2 border-gray bg-lightgray border rounded-full py-2 px-3 lg:text-sm md:text-sm sm:text-xs xs:text-xs mx-auto   hover:bg-SlateBlue hover:text-white hover:bg-primary-500 hover:shadow-lg hover:shadow-primary-500/50"
+                      {nowPlayingContentVisible && (
+                        <td
+                          className="p-2 text-center "
+                          style={{ wordBreak: "break-all" }}
                         >
-                          <p className="line-clamp-3">{screen.assetName}</p>
-                          <AiOutlineCloudUpload className="min-h-[1rem] min-w-[1rem]" />
-                        </div>
-                        {showAssetModal && (
-                          <ShowAssetModal
-                            handleAssetAdd={handleAssetAdd}
-                            handleAssetUpdate={handleAssetUpdate}
-                            setSelectedComposition={setSelectedComposition}
-                            handleAppsAdd={handleAppsAdd}
-                            popupActiveTab={popupActiveTab}
-                            setAssetPreviewPopup={setAssetPreviewPopup}
-                            setPopupActiveTab={setPopupActiveTab}
-                            setShowAssetModal={setShowAssetModal}
-                            assetPreviewPopup={assetPreviewPopup}
-                            assetPreview={assetPreview}
-                            selectedComposition={selectedComposition}
-                            selectedTextScroll={selectedTextScroll}
-                            selectedYoutube={selectedYoutube}
-                            selectedAsset={selectedAsset}
-                          />
-                        )}
-                      </td>
-                    )}
-                    {currScheduleContentVisible && (
-                      <td className="break-words	w-[150px] p-2 text-center">
-                        {screen.scheduleName == "" ? (
-                          <button
-                            onClick={() => {
-                              setShowScheduleModal(true);
-                              setScheduleScreenID(screen.screenID);
+                          <div
+                            onClick={(e) => {
+                              setAssetScreenID(screen.screenID);
+                              setShowAssetModal(true);
+                              setSelectedAsset({
+                                ...selectedAsset,
+                                assetName: e.target.value,
+                              });
+                              setSelectedAsset(screen?.assetName);
                             }}
+                            title={screen?.assetName}
+                            className="flex items-center justify-between gap-2 border-gray bg-lightgray border rounded-full py-2 px-3 lg:text-sm md:text-sm sm:text-xs xs:text-xs mx-auto   hover:bg-SlateBlue hover:text-white hover:bg-primary-500 hover:shadow-lg hover:shadow-primary-500/50"
                           >
-                            Set a schedule
-                          </button>
-                        ) : (
-                          `${screen.scheduleName} Till
+                            <p className="line-clamp-3">{screen.assetName}</p>
+                            <AiOutlineCloudUpload className="min-h-[1rem] min-w-[1rem]" />
+                          </div>
+                          {showAssetModal && (
+                            <ShowAssetModal
+                              handleAssetAdd={handleAssetAdd}
+                              handleAssetUpdate={handleAssetUpdate}
+                              setSelectedComposition={setSelectedComposition}
+                              handleAppsAdd={handleAppsAdd}
+                              popupActiveTab={popupActiveTab}
+                              setAssetPreviewPopup={setAssetPreviewPopup}
+                              setPopupActiveTab={setPopupActiveTab}
+                              setShowAssetModal={setShowAssetModal}
+                              assetPreviewPopup={assetPreviewPopup}
+                              assetPreview={assetPreview}
+                              selectedComposition={selectedComposition}
+                              selectedTextScroll={selectedTextScroll}
+                              selectedYoutube={selectedYoutube}
+                              selectedAsset={selectedAsset}
+                            />
+                          )}
+                        </td>
+                      )}
+                      {currScheduleContentVisible && (
+                        <td className="break-words	w-[150px] p-2 text-center">
+                          {screen.scheduleName == "" ? (
+                            <button
+                              onClick={() => {
+                                setShowScheduleModal(true);
+                                setScheduleScreenID(screen.screenID);
+                              }}
+                            >
+                              Set a schedule
+                            </button>
+                          ) : (
+                            `${screen.scheduleName} Till
                         ${moment(screen.endDate).format("YYYY-MM-DD hh:mm")}`
-                        )}
+                          )}
 
-                        {showScheduleModal && (
-                          <div className="bg-black bg-opacity-50 justify-center items-center flex overflow-x-hidden overflow-y-auto fixed inset-0 z-50 outline-none focus:outline-none">
-                            <div className="w-auto my-6 mx-auto lg:max-w-6xl md:max-w-xl sm:max-w-sm xs:max-w-xs">
-                              <div className="border-0 rounded-lg shadow-lg relative flex flex-col w-full bg-white outline-none focus:outline-none">
-                                <div className="flex items-start justify-between p-4 px-6 border-b border-[#A7AFB7] rounded-t text-black">
-                                  <div className="flex items-center">
-                                    <h3 className="lg:text-xl md:text-lg sm:text-base xs:text-sm font-medium">
-                                      Set Schedule
-                                    </h3>
+                          {showScheduleModal && (
+                            <div className="bg-black bg-opacity-50 justify-center items-center flex overflow-x-hidden overflow-y-auto fixed inset-0 z-50 outline-none focus:outline-none">
+                              <div className="w-auto my-6 mx-auto lg:max-w-6xl md:max-w-xl sm:max-w-sm xs:max-w-xs">
+                                <div className="border-0 rounded-lg shadow-lg relative flex flex-col w-full bg-white outline-none focus:outline-none">
+                                  <div className="flex items-start justify-between p-4 px-6 border-b border-[#A7AFB7] rounded-t text-black">
+                                    <div className="flex items-center">
+                                      <h3 className="lg:text-xl md:text-lg sm:text-base xs:text-sm font-medium">
+                                        Set Schedule
+                                      </h3>
+                                    </div>
+                                    <button
+                                      className="p-1 text-xl"
+                                      onClick={() =>
+                                        setShowScheduleModal(false)
+                                      }
+                                    >
+                                      <AiOutlineCloseCircle className="text-2xl" />
+                                    </button>
                                   </div>
-                                  <button
-                                    className="p-1 text-xl"
-                                    onClick={() =>
-                                      setShowScheduleModal(false)
-                                    }
-                                  >
-                                    <AiOutlineCloseCircle className="text-2xl" />
-                                  </button>
-                                </div>
-                                <div className="overflow-x-auto mt-8 px-5">
-                                  <table
-                                    className="w-full  lg:table-fixed md:table-auto sm:table-auto xs:table-auto"
-                                    cellPadding={20}
-                                  >
-                                    <thead>
-                                      <tr className="items-center border-b border-b-[#E4E6FF] table-head-bg text-left">
-                                        <th className="font-medium text-[14px]">
-                                          <div className="flex items-center">
-                                            <TbCalendarTime className="mr-2 text-xl" />
-                                            Schedule Name
-                                          </div>
-                                        </th>
-                                        <th className="font-medium text-[14px]">
-                                          <div className="flex items-center">
-                                            <TbCalendarTime className="mr-2 text-xl" />
-                                            Time Zones
-                                          </div>
-                                        </th>
-                                        <th className="font-medium text-[14px]">
-                                          <div className=" flex  items-center justify-center mx-auto">
-                                            <VscCalendar className="mr-2 text-xl" />
-                                            Date Added
-                                          </div>
-                                        </th>
-                                        <th className="font-medium text-[14px]">
-                                          <div className=" flex  items-center justify-center mx-auto">
-                                            <TbCalendarStats className="mr-2 text-xl" />
-                                            start date
-                                          </div>
-                                        </th>
-                                        <th className="font-medium text-[14px]">
-                                          <div className=" flex  items-center justify-center mx-auto">
-                                            <TbCalendarStats className="mr-2 text-xl" />
-                                            End date
-                                          </div>
-                                        </th>
-                                        <th className="font-medium text-[14px]">
-                                          <div className=" flex  items-center justify-center mx-auto">
-                                            <RiComputerLine className="mr-2 text-xl" />
-                                            screens Assigned
-                                          </div>
-                                        </th>
-                                        <th className="font-medium text-[14px]">
-                                          <div className="flex  items-center justify-center mx-auto">
-                                            <BsTags className="mr-2 text-xl" />
-                                            Tags
-                                          </div>
-                                        </th>
-                                        <th></th>
-                                      </tr>
-                                    </thead>
-                                    <tbody>
-                                      {schedules.map((schedule) => (
-                                        <tr
-                                          className="mt-7 bg-white rounded-lg  font-normal text-[14px] text-[#5E5E5E] border-b border-lightgray shadow-sm px-5 py-2"
-                                          key={schedule.scheduleId}
-                                        >
-                                          <td className="flex items-center ">
-                                            <input
-                                              type="checkbox"
-                                              className="mr-3"
-                                              onChange={() =>
-                                                handleScheduleAdd(schedule)
-                                              }
-                                            />
-                                            <div>
-                                              <div>
-                                                {schedule.scheduleName}
-                                              </div>
+                                  <div className="overflow-x-auto mt-8 px-5">
+                                    <table
+                                      className="w-full  lg:table-fixed md:table-auto sm:table-auto xs:table-auto"
+                                      cellPadding={20}
+                                    >
+                                      <thead>
+                                        <tr className="items-center border-b border-b-[#E4E6FF] table-head-bg text-left">
+                                          <th className="font-medium text-[14px]">
+                                            <div className="flex items-center">
+                                              <TbCalendarTime className="mr-2 text-xl" />
+                                              Schedule Name
                                             </div>
-                                          </td>
-                                          <td className="text-center">
-                                            {schedule.timeZoneName}
-                                          </td>
-                                          <td className="text-center">
-                                            {moment(
-                                              schedule.createdDate
-                                            ).format("YYYY-MM-DD hh:mm")}
-                                          </td>
-                                          <td className="text-center">
-                                            {moment(
-                                              schedule.startDate
-                                            ).format("YYYY-MM-DD hh:mm")}
-                                          </td>
-
-                                          <td className="text-center">
-                                            {moment(schedule.endDate).format(
-                                              "YYYY-MM-DD hh:mm"
-                                            )}
-                                          </td>
-                                          <td className="p-2 text-center">
-                                            {schedule.screenAssigned}
-                                          </td>
-                                          <td className="p-2 text-center">
-                                            {schedule.tags}
-                                          </td>
-                                          <td className="text-center">
-                                            <Link to="/myschedule">
-                                              <button className="ml-3 relative">
-                                                <HiDotsVertical />
-                                              </button>
-                                            </Link>
-                                          </td>
+                                          </th>
+                                          <th className="font-medium text-[14px]">
+                                            <div className="flex items-center">
+                                              <TbCalendarTime className="mr-2 text-xl" />
+                                              Time Zones
+                                            </div>
+                                          </th>
+                                          <th className="font-medium text-[14px]">
+                                            <div className=" flex  items-center justify-center mx-auto">
+                                              <VscCalendar className="mr-2 text-xl" />
+                                              Date Added
+                                            </div>
+                                          </th>
+                                          <th className="font-medium text-[14px]">
+                                            <div className=" flex  items-center justify-center mx-auto">
+                                              <TbCalendarStats className="mr-2 text-xl" />
+                                              start date
+                                            </div>
+                                          </th>
+                                          <th className="font-medium text-[14px]">
+                                            <div className=" flex  items-center justify-center mx-auto">
+                                              <TbCalendarStats className="mr-2 text-xl" />
+                                              End date
+                                            </div>
+                                          </th>
+                                          <th className="font-medium text-[14px]">
+                                            <div className=" flex  items-center justify-center mx-auto">
+                                              <RiComputerLine className="mr-2 text-xl" />
+                                              screens Assigned
+                                            </div>
+                                          </th>
+                                          <th className="font-medium text-[14px]">
+                                            <div className="flex  items-center justify-center mx-auto">
+                                              <BsTags className="mr-2 text-xl" />
+                                              Tags
+                                            </div>
+                                          </th>
+                                          <th></th>
                                         </tr>
-                                      ))}
-                                    </tbody>
-                                  </table>
-                                </div>
+                                      </thead>
+                                      <tbody>
+                                        {schedules.map((schedule) => (
+                                          <tr
+                                            className="mt-7 bg-white rounded-lg  font-normal text-[14px] text-[#5E5E5E] border-b border-lightgray shadow-sm px-5 py-2"
+                                            key={schedule.scheduleId}
+                                          >
+                                            <td className="flex items-center ">
+                                              <input
+                                                type="checkbox"
+                                                className="mr-3"
+                                                onChange={() =>
+                                                  handleScheduleAdd(schedule)
+                                                }
+                                              />
+                                              <div>
+                                                <div>
+                                                  {schedule.scheduleName}
+                                                </div>
+                                              </div>
+                                            </td>
+                                            <td className="text-center">
+                                              {schedule.timeZoneName}
+                                            </td>
+                                            <td className="text-center">
+                                              {moment(
+                                                schedule.createdDate
+                                              ).format("YYYY-MM-DD hh:mm")}
+                                            </td>
+                                            <td className="text-center">
+                                              {moment(
+                                                schedule.startDate
+                                              ).format("YYYY-MM-DD hh:mm")}
+                                            </td>
 
-                                <div className="py-4 flex justify-center">
-                                  <button
-                                    onClick={() => {
-                                      setShowScheduleModal(false);
-                                      handleScheduleUpdate(screen.screenID);
-                                    }}
-                                    className="bg-SlateBlue border-primary px-5 py-2 rounded-full ml-3"
-                                  >
-                                    Save
-                                  </button>
+                                            <td className="text-center">
+                                              {moment(schedule.endDate).format(
+                                                "YYYY-MM-DD hh:mm"
+                                              )}
+                                            </td>
+                                            <td className="p-2 text-center">
+                                              {schedule.screenAssigned}
+                                            </td>
+                                            <td className="p-2 text-center">
+                                              {schedule.tags}
+                                            </td>
+                                            <td className="text-center">
+                                              <Link to="/myschedule">
+                                                <button className="ml-3 relative">
+                                                  <HiDotsVertical />
+                                                </button>
+                                              </Link>
+                                            </td>
+                                          </tr>
+                                        ))}
+                                      </tbody>
+                                    </table>
+                                  </div>
+
+                                  <div className="py-4 flex justify-center">
+                                    <button
+                                      onClick={() => {
+                                        setShowScheduleModal(false);
+                                        handleScheduleUpdate(screen.screenID);
+                                      }}
+                                      className="bg-SlateBlue border-primary px-5 py-2 rounded-full ml-3"
+                                    >
+                                      Save
+                                    </button>
+                                  </div>
                                 </div>
                               </div>
                             </div>
-                          </div>
-                        )}
-                      </td>
-                    )}
+                          )}
+                        </td>
+                      )}
 
-                    {tagsContentVisible && (
-                      <td 
-                      title={screen?.tags && screen?.tags}
-                      className="p-2 text-center flex flex-wrap items-center justify-center gap-2 mt-6 break-all">
-                        {(screen?.tags === "" || screen?.tags === null) && (
-                          <span>
-                            <AiOutlinePlusCircle
-                              size={30}
-                              className="mx-auto cursor-pointer"
+                      {tagsContentVisible && (
+                        <td
+                          title={screen?.tags && screen?.tags}
+                          className="p-2 text-center flex flex-wrap items-center justify-center gap-2 mt-6 break-all"
+                        >
+                          {(screen?.tags === "" || screen?.tags === null) && (
+                            <span>
+                              <AiOutlinePlusCircle
+                                size={30}
+                                className="mx-auto cursor-pointer"
+                                onClick={() => {
+                                  setShowTagModal(true);
+                                  screen.tags === "" || screen?.tags === null
+                                    ? setTags([])
+                                    : setTags(screen?.tags?.split(","));
+                                  setTagUpdateScreeen(screen);
+                                }}
+                              />
+                            </span>
+                          )}
+
+                          {screen?.tags !== null
+                            ? screen.tags
+                                .split(",")
+                                .slice(
+                                  0,
+                                  screen.tags.split(",").length > 2
+                                    ? 3
+                                    : screen.tags.split(",").length
+                                )
+                                .map((text) => {
+                                  if (text.toString().length > 10) {
+                                    return text
+                                      .split("")
+                                      .slice(0, 10)
+                                      .concat("...")
+                                      .join("");
+                                  }
+                                  return text;
+                                })
+                                .join(",")
+                            : ""}
+                          {screen?.tags !== "" && screen?.tags !== null && (
+                            <MdOutlineModeEdit
                               onClick={() => {
                                 setShowTagModal(true);
                                 screen.tags === "" || screen?.tags === null
@@ -1947,98 +1990,65 @@ const Screens = ({ sidebarOpen, setSidebarOpen }) => {
                                   : setTags(screen?.tags?.split(","));
                                 setTagUpdateScreeen(screen);
                               }}
+                              className="w-5 h-5 cursor-pointer"
                             />
-                          </span>
-                        )}
+                          )}
 
-                        {screen?.tags !== null
-                          ? screen.tags
-                              .split(",")
-                              .slice(
-                                0,
-                                screen.tags.split(",").length > 2
-                                  ? 3
-                                  : screen.tags.split(",").length
-                              ).map((text) => {
-                                if (text.toString().length > 10) {
-                                  return text
-                                    .split("")
-                                    .slice(0, 10)
-                                    .concat("...")
-                                    .join("");
-                                }
-                                return text;
-                              })
-                              .join(",")
-                          : ""}
-                        {screen?.tags !== "" && screen?.tags !== null && (
-                          <MdOutlineModeEdit
-                            onClick={() => {
-                              setShowTagModal(true);
-                              screen.tags === "" || screen?.tags === null
-                                ? setTags([])
-                                : setTags(screen?.tags?.split(","));
-                              setTagUpdateScreeen(screen);
-                            }}
-                            className="w-5 h-5 cursor-pointer"
-                          />
-                        )}
+                          {/* add or edit tag modal */}
+                          {showTagModal && (
+                            <AddOrEditTagPopup
+                              setShowTagModal={setShowTagModal}
+                              tags={tags}
+                              setTags={setTags}
+                              handleTagsUpdate={handleTagsUpdate}
+                              from="screen"
+                              setTagUpdateScreeen={setTagUpdateScreeen}
+                            />
+                          )}
+                        </td>
+                      )}
 
-                        {/* add or edit tag modal */}
-                        {showTagModal && (
-                          <AddOrEditTagPopup
-                            setShowTagModal={setShowTagModal}
-                            tags={tags}
-                            setTags={setTags}
-                            handleTagsUpdate={handleTagsUpdate}
-                            from="screen"
-                            setTagUpdateScreeen={setTagUpdateScreeen}
-                          />
-                        )}
-                      </td>
-                    )}
-
-                    <td className="p-2 text-center relative">
-                      <div className="relative">
-                        <button
-                          className="ml-3 relative"
-                          onClick={() => handleScreenClick(screen.screenID)}
-                        >
-                          <HiDotsVertical />
-                        </button>
-                        {/* action popup start */}
-                        {showActionBox[screen.screenID] && (
-                          <div
-                            ref={showActionModalRef}
-                            className="scheduleAction"
+                      <td className="p-2 text-center relative">
+                        <div className="relative">
+                          <button
+                            className="ml-3 relative"
+                            onClick={() => handleScreenClick(screen.screenID)}
                           >
-                            <div className="my-1">
-                              <Link
-                                to={`/screensplayer?screenID=${screen.screenID}`}
-                              >
-                                <button className="text-sm">
-                                  Edit Screen
-                                </button>
-                              </Link>
-                            </div>
+                            <HiDotsVertical />
+                          </button>
+                          {/* action popup start */}
+                          {showActionBox[screen.screenID] && (
+                            <div
+                              ref={showActionModalRef}
+                              className="scheduleAction"
+                            >
+                              <div className="my-1">
+                                <Link
+                                  to={`/screensplayer?screenID=${screen.screenID}`}
+                                >
+                                  <button className="text-sm">
+                                    Edit Screen
+                                  </button>
+                                </Link>
+                              </div>
 
-                            <div className="mb-1 border border-[#F2F0F9]"></div>
-                            <div className=" mb-1 text-[#D30000]">
-                              <button
-                                onClick={() =>
-                                  handelDeleteScreen(screen.screenID)
-                                }
-                                className="text-sm text-left"
-                              >
-                                Delete
-                              </button>
+                              <div className="mb-1 border border-[#F2F0F9]"></div>
+                              <div className=" mb-1 text-[#D30000]">
+                                <button
+                                  onClick={() =>
+                                    handelDeleteScreen(screen.screenID)
+                                  }
+                                  className="text-sm text-left"
+                                >
+                                  Delete
+                                </button>
+                              </div>
                             </div>
-                          </div>
-                        )}
-                      </div>
-                    </td>
-                  </tr>
-                </tbody>
+                          )}
+                        </div>
+                      </td>
+                    </tr>
+                  </tbody>
                 ))
               )}
             </table>
